Fixed support for streaming in stdout events by manually adding a eof event
The problem with not having an explicit eof event is that it becomes
hard to understand to communicate stop reading vs. there’s no more to
read. The only other thing I could figure out is to use something like a
timeout but even that seems wrong.

Readline implementations block until they read a line or presumably
until a channel is closed or until they read some message from the
channel. Because we want to run more code after an event, closing the
program and or channel doesn’t work in this case. If we don’t send
anything then readline blocks forever. So all we can really do is pass
an EOF explicitly to communicate an EOF. This is basically the same way
server-sent events work because presumably they have the same kind of
problem.